+++ Sibusten.Philomena.Downloader.Cmd/Commands/Booru/Arguments/BooruListCommandArgs.cs
@@ -1,0 +1,7 @@
+namespace Sibusten.Philomena.Downloader.Cmd.Commands.Booru.Arguments
+{
+    public class BooruListCommandArgs
+    {
+        public bool Verbose { get; set; } = false;
+    }
+}