+++ Sibusten.Philomena.Downloader.Cmd/Commands/Booru/Arguments/BooruRemoveCommandArgs.cs
@@ -1,0 +1,7 @@
+namespace Sibusten.Philomena.Downloader.Cmd.Commands.Booru.Arguments
+{
+    public class BooruRemoveCommandArgs
+    {
+        public string Name { get; set; } = "";
+    }
+}