+++ Sibusten.Philomena.Downloader.Cmd/Commands/Booru/Arguments/BooruRenameCommandArgs.cs
@@ -1,0 +1,8 @@
+namespace Sibusten.Philomena.Downloader.Cmd.Commands.Booru.Arguments
+{
+    public class BooruRenameCommandArgs
+    {
+        public string From { get; set; } = "";
+        public string To { get; set; } = "";
+    }
+}