+++ Sibusten.Philomena.Downloader.Cmd/Commands/Booru/Arguments/BooruUpdateCommandArgs.cs
@@ -1,0 +1,9 @@
+namespace Sibusten.Philomena.Downloader.Cmd.Commands.Booru.Arguments
+{
+    public class BooruUpdateCommandArgs
+    {
+        public string Name { get; set; } = "";
+        public string? BaseUrl { get; set; }
+        public string? ApiKey { get; set; }
+    }
+}